Append a newline to redered templates
For whatever reason, the jinja render does not seem to like trailing
newlines - which means it generates files with noeol. vim helpfully
fixes that for people, which means that generated files that are then
edited get an extra newline added in addition to the first edit.

Simple fix - add newline.