Support reading boolean config values
If the default value of an entry in `cookiecutter.json` is a boolean
then preserve further user input as boolean using read_user_yes_no.

For example, the json:

    {
        'is_foo': true
    }

would result in a prompt like

    is_foo [True]:

Then if a user input "false", "False", "n", etc it would result in a
False value in python.