added CookIeCutter Automated DRF Angular (CICADA)
CICADA uses Cookiecutter and a custom script to automatically populate an AngularJS frontend that connects to your Django REST backend. This template makes some assumptions about your backend, namely that it is a Django DRF-based backend with JWT authentication. Such a template for a backend can be easily created with the cc-automated-drf-template.