__target__ setting for freestanding files
If __target__ is set in cookiecutter.json then
files are placed at this location within the directory
tree rather than creating a new directory.

Pre-existing files are unaffected. A common case
would be __target__ to '.' . The source directory
in the template can be any cookiecutter template expression,
but this is ignored. A value like {{cookiecutter.files}}
is suggested.