Added option to strip template directory.
The following commit implements a new feature: stripping the template directory.
When you create your project based on cookiecutter template, it's using the first
jinja directory as template directory and render it.
Even if you specify an output directory it will always create a sub-rendered-directory.

The strip option skip this first template directory and render the boilerplate in
output directory.