Adds new command: cookiecutter installed
This lists locally available templates, for example:

```
user$ cookiecutter installed
7 installed templates:
 * cookiecutter-bottle
 * cookiecutter-pyramid-talk-python-starter
 * cookiecutter-sublime-text-3-plugin
 * cookiecutter-template
 * puppet-openstack-cookiecutter
 * pyramid-cookiecutter-starter
 * python-macon-template
```

I was not able to integrate it into click without breaking the existing functionality. Someone who knows click better than me please adjust this if possible.

I’d like to see just `cookiecutter —installed` as the command (and it would need to list in help) but I only get missing template and all attempts to add it as an option met with my lack of exepreience with click.