Update unit tests for Click 7.0
Fixed two unit tests that fail with Click 7.0 (see #1109 for details).

Fixes #1109.

The tests now pass with Click 7.0, but will fail with older versions of Click due to small differences in usage error messages. Do the tests need to pass with all supported versions of Click (>=5.0) or is it sufficient to have them pass with the latest supported version of Click?