Fix for deprecationwarning on python 3.6
When running tests on travis with python 3.6, I got the following
deprecationwarning in cookiecutter/repository.py:19:

    DeprecationWarning: Flags not at the start of the expression '\n(?x)\n((((git|hg)\\+)' (truncated)
    """)

Python 3.6 prefers the flags right at the start, so it doesn't allow a newline
in front of it.
I fixed it by adding the flag at the end, which is more clear/verbose in any
case. Which is fitting as the flag is `re.VERBOSE` :-)