Replace hardcoded template listing
Replaces the hardcoded listing of templates with generic categories pointing to github topic queries.

This should cut the big number of listing updates and also provide an unbiased listing method, as github default listing is based on projects popularity (stars).

We also need to send en email that warns template publishers to update their topics on github. If possible we should try to send this email before merging the change.