json schema changed
the structure of the schema was slightly changed to be more nested, so
it looks better when displayed as yaml.
also, an issue with additional properties was fixed: additional
unexpected fields are not allowed anymore.

this change will break a lot of tests, because only the schema and the
tests that directly depend on the schema were changed.