Make read_user_dict compatible with click 8.x
The `read_user_dict` function uses a "default" sentinel instead of the actual
default value. Being a JSON dict, the latter would often be hard to type.

Under click 8.x, the default value for `click.prompt` is passed to the
`read_proc` callback. We use this callback to load JSON from the user input, and
this would choke on an input like "default" (without quotes). Therefore, change
the callback to return the default value when it receives the "default"
sentinel.

Under click 7.x (which is our minimum version), the default value for
`click.prompt` is returned as-is. Therefore, continue to handle the case where
`click.prompt` returns "default" instead of the actual default value, but only
if we're actually running under click 7.x.

(Checking for click 7.x is only done for clarity. Under click 8.x,
`click.prompt` would never return "default", even if a user entered it as a
valid JSON string. This is because our callback requires a dict, not a string.)