fix so dict val parsing isn't deferred
Currently, you cannot reference another value within the cookiecutter.json file if the referenced value is a dict, because those values are deferred to a second pass for parsing.  The cookiecutter.json is parsed as an OrderedDict, therefore there is no benefit to do multiple passes as it only prevents the use of some vars (dicts) in later values.  If a val depends on another value within the json file, it only needs to be put earlier/higher within the file and then can be used anywhere later in the file.