fix: FileExistsError when using relative path
When using a local cookiecutter template ('cookiecutter .')
that contains a pre_prompt script, cookiecutter will exit
with a FileExistsError. The stack trace hints that the temporary
directory already exists and cannot be created by shutils.copytree.

By converting repo_dir into a Path object and calling resolve
to get the absolute path the actual name of the directory is used
instead of '.'.