fix: modifying start and end variable strings
Cookiecutter supports using _jinja2_env_vars to modify the jinja environment.
However when modifying variable_start_string or variable_end_string it will not work
because we are hard-coding the start and end strings for the main directory creation.

This change leverages the _jinja2_env_vars to search for the main directory, to allow
this customization of variable_start_string and variable_end_string.

I also remove the  check because the same logic
is used to find the dir, which is therefore redundant.