fix: convert strings to booleans for boolean config variables (#1973)
Fixes a bug where the override value provided via command line wasn't being properly converted to its boolean counterpart.

Since override values from the command line are interpreted as strings, we need to convert the string value to a boolean if the config variable is a boolean to begin with.